simplify remove lib.rs
